backend/pyproject.toml
[project]
name = "pitwall-tools"
version = "0.1.0"
description = "PitWall Live - Real-Time AI Race Engineer Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.1",
    "redis>=5.0.3",
    "fastf1>=3.4.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
